Rename EventInformation class to AtlasG4EventUserInfo
The `EventInformation` class inherits from the `G4VUserEventInformation` class.
It is used to pass around additional information about the truth particle
currently being processed.
However, the name is very similar to the `(xAOD::)EventInfo` object used
elsewhere in Athena. In order to avoid this clash this commit renames the
`EventInformation` class to `AtlasG4EventUserInfo`.