Avoid duplicate decisions for muons in muon triggers
The precision muon trigger hypo loops first on the previous decisions, then gets the event view for each decision, and finally retrievs the muons from the view. In the case where we have an input maker that merges based on feature, we run into the problem of having one decision per unique muon, which means we have N decisions for a given view, and so end up creating N new decisions per muon. To avoid this, I've introduced a check that the muon corresponds to the muon from the previous decision. Since this is only a problem for combined muons, we first check if the extrapolated muon spectrometer track for the combined muons matches the extrapolated track from the previous step's SA muons. In case there is no extrapolated track (which is the case for muons reconstructed by the InsideOut reconstruction), we check the dR between the combined muons and the previous step's muons. This makes sure we don't create new decisions multiple times for the same muon (which is reflected in the reference update), and avoids warnings of the form:

`WARNING Found 'feature' by multiple paths, but the links are coming from different places in the navigation graph`

as discussed in ATR-21548.