MadGraphControl SUSY: Allowing multiple input LHE files
Some of the manipulations in the SUSY post-exec assume that we have a
tarball as input, which happens when a single input file is used. When
multiple input files come in, we have an LHE file as input, and in that
case we need to handle things a bit differently (mostly avoid all of the
tarball manipulations and instead treat the thing as a text file
directly).