PileUpTools: Protect against multiple attempts to build EventInfo.
PileUpMergeSvc::getPileUpEvent will try to construct an EventInfo object
if it doesn't exist.  But if this gets called multiple times for the same 
slot, then we can make multiple EventInfo objects and eventually get
errors due to trying to record duplicates in SG.  Add a set of
slot-specific mutexes to protect against this.