Migrate GX2Fitter to in-class Gaudi properties
As far as I know, using in-class Gaudi properties is the preferred and
modern way of declaring properties for a tool. The GlobalChi2Fitter is
currently still using the older in-constructor style properties, and
this commit changes this to the new style. Besides being more modern,
idiomatic Gaudi, it also means the properties, their config-level
names, and their default values are now in a single place is which a
lot easier to wrap the head around. The only property not migrated was
a service handle, because of some runtime parameters being passed to
its declaration statement.