Remove const_cast from MuonErrorOptimisationTool
As the track passed into the tool is always non-const, there was no reason for the tracks that are being passed around inside the tool, or the MuonRefitTool that it calls, to be const.
The const_casts compensated for this not-really-constness: removing it means that they can be removed as well.
Additionally, I took advantage of this change to switch to using unique_ptr for the tracks, and did some minor cleanup in the MuonRefitTool.