AthenaKernel: Additional DataBucketBase::cast overload.
There have been two overloads for DataBucketBase::cast, one taking
a clid and one taking a type_info.  To cast from a DataProxy, we've
been using the type_info version.  However, doing it with a clid
can be a bit faster, so we'd like to use that.  There is a problem,
though, since xAODRootAccess also implements a DataBucket, but this
version only supports casting via type_info.

So add a third overload for cast that takes both a clid and a type_info
and give the callee the opportunity to choose which one to use.