RootUtils: Updates to help with root 6.22.
The performance hacks in PyROOTTreePatch are no longer needed.
Remove them, and in so doing we can remove the Converter code that
was copied from PyROOT.  Retain for now the functionality of
pythonizing SetNotify and TBranch::SetAddress.

Clean up Utility to remove things that aren't needed.
Add some #ifdefs for pyroot 6.20 / 6.22 compatibility.

Fix the TTreePatch test.