Minor style changes to IMatEffUpdator and related
This is a pretty meaningless commit by itself, but its purpose is to
facilitate a future commit in which we change the function signatures
of these methods. Since the current formatting is dependent on the
return type, it's not possible to: 1) change the return types of the
methods; 2) keep the formatting consistent; and 3) keep unrelated
changes out of the next commit. To fix this, this commit first changes
the formatting to be independent of the return types, so we can change
them without having to change more formatting.