Use smart pointers in IMaterialEffectsUpdator
In the existing code, the IMaterialEffectsUpdator methods return raw
pointers to TrackParameters. This is most likely an artifact of the
bad old days where we didn't have proper pointer ownership mechanics
built into the C++ standard, but we now live in more enlightened times
and returning raw pointers from methods which assume the caller to
assume ownership of the returned value simply won't do anymore. We
need to embed as much semantic meaning about ownership in the function
signatures as possible!

In merge request !35528, the interface was already modified such that
it now guarantees non-aliasing return values, but the ownership
semantics were not there. This commit adds those semantics such that
it is much clearer to the user of the interface that the caller should
assume ownership of the returned value. After all, they don't have
much choice: the compiler will scream at them unless they explicitly
release the unique pointer.

Relates to ATLASRECTS-5614. This commit should partially resolve that
memory leak issue. Hopefully.