Fix GX2F reference parameter cloning memleak
This commit fixes a minor mistake in !35388 that may lead to a small
but nonetheless important memory leak. During the move away from the
old memory management inside the fitter, I accidentally added an extra
and unnecessary clone of a track's reference parameters which would
result in small memory leak. This commit gets rids of this vestige of
the old memory model by modifying the getters, setters, and internal
representation of a track's reference parameters and replacing them by
the standard smart pointer ownership semantics. The extraneous clone
is removed as a direct result of this.

Related to ATLASRECTS-5614.