Alternative path to truth link for small collections
When running on TRUTH3 collections, the truth particle links from reco
objects show up as TruthLink rather than truthParticleLink (because of
the naming conventions in the truth relink code in
DerivationFrameworkMCTruth). This updates the truth helpers to look in
the standard location, and if the links cannot be resolved for any
reason to fall back to look in that alternative location. Resolves the
latest question in ATLASG-1502.