TRT_RawDataByteStreamCnv: threading fixes
Rework to avoid calling naughtyRetrieve.

Remove ATLAS_NOT_THREAD_SAFE annotations --- just need to avoid a (needless!)
AttributeList copy.

Make sure to protect TRTRawDataProviderTool::m_cache with a lock.
Not safe to omit it.

Note that the original code was sometimes unpacking the same channel
multiple times into the same collection.  This behavior is fixed
by the change, but it results in RootComp differences for a couple
TriggerTest ART tests.