Refactor GlobalChi2Fitter energy loss updator
Here, again, we have some code that is duplicated for downstream and
upstream track states, which makes the code less clear than it can be.
We lift the code related to updating track states for material effect
energy loss into a new, dedicated method with more sensible memory
ownership semantics.