remove package CPAnalysisExamples
This package is outdated in many ways, and mostly contains examples
for writing CP tools that we cooked up during LS1.  Besides providing
outdated examples of how to do things it also includes a conflicting
`IJetCalibrationTool` implementation (possibly others as well),
meaning it can't even be included in any release that also contains
the regular CP tool packages.

Now all of this could be fixed, but it would probably be easier to
rewrite the examples from scratch instead of trying to fix this up.
And I am not at all convinced that this is even a worthwhile
endeavour.  These examples were provided to give CP tool developer an
idea of how to develop the CP tools in the first place and that is
done now.  For the question of how to use CP tools we would probably
point users to the common CP algorithms these days.  And for either
one there is plenty of production level code available that people can
look at (though that is not a full replacement for educational
examples).