SGTools; Allow marking keys as transient to avoid hash collisions.
When we record a new object to SG, we get an error if the key+clid
hashes to an already-existing value.  Allow marking a SG string key
as `transient' by prefixing it with an underscore.  In this case,
we don't fail on a hash collision, but instead assign this key+clid
an unused hash.  This is only allowed for transient SG keys,
since the hash used will not correspond to the result of the hash
function for these cases, which could be confusing for persistent data.