Use same HypoAlg for ms-only and standard muon triggers
Switching the msonly muon triggers to use exactly the same SA muon trigger sequence as other chains. This is to avoid having two different hypo algs attach the same feature twice in the navigation (which can lead to retrieving the incorrect features from the navigation later on in the monitoring).