Temporary workaround to fix issue in the TDT feature retrieval
In the TDT, when requesting features, it *also* returns objects with
matching names, such that when requesting lastFeatureOfType, if it finds
an object which matches before the exact match, it will stop.

This will be fixed in the TDT feature access, but until the we add this
temporary  workaround to request *all* the features of the type, and then
reject those that do not match explicitly. Will probably revert the code
once the issue in the TDT access is resolved, so leave the original
selection here, but commented