RootStorageSvc: Work around SetAddress change in root 6.22.
As of root 6.22, calling SetAddress with nullptr may not work as expected
if the address had previously been set to something non-null.
We were doing this if we wanted root to create a temporary empty
object for output, but this doesn't work properly with 6.22.
Instead, create a temporary object ourselves.