Add a full scan test, add additional panels to the plotting, fix the incorrect selection of histograms
An extra validation test running the fullscan tests is added - running from a jet chain, rather than
a bjet chain (will need a jet performance test at some point)

In addition, it adds extra panels for the plotting showing the hit multiplicities and the raw distributions.

Finally, a fix to the configuration of the histogram selection in the FTF monitoring has been added.