Fixing check for compatibility between release and jet collection name
The compatibility check failed for release numbers above 99 due to
simple string comparison (did not properly compare two-digits release
numbers with three-digits needed for releases 21.2.100 and above).
This caused runtime error incorrectly stating that the configured
release was older than 21.2.72.0. This patch corrects for that by
instead comparing the release numbers as integers. An informative
message stating the release number read from the sample meta data
was also added to the INFO output.