Add tauRec flag for minimum seed jet pt
Hello,

This MR is introducing a new tauRec flag to control the minimum pt of the seed jet, instead of a currently hardcoded 10 GeV cut.
This will be useful to study the impact of that cut on the tau reconstruction efficiency.
Some minor left-over pieces have been cleaned up.

Cheers,
Bertrand