TriggerJobOpts: Updates for magnetic field settings (ATR-21277)
Update various job options to work with the new magnetic field
conditions algorithm and retire the use of the AtlasFieldSvc. Currently
we need to run with
```
AtlasFieldMapCondAlg.UseMapsFromCOOL = False
```
because loading the field map names from COOL does not work during the
start transition. Alternatively we could read the maps from COOL on
the first event, but this would entail a memory penalty when forking.
Map names never/rarely change. But if they do and if we haven't changed
to using them from COOL we need to make sure they are updated also in
the job option parameters.