Use TSOS types instead of TS types internally
In the existing architecture, track state types for the
GlobalChi2Fitter's internal representation of a track are captured
using the TrackState::TrackStateType enumeration class. This seems to
be a relic of times long past, but it doesn't quite capture the
semantics we need to properly represent track state types (although it
is close). Firstly, it doesn't allow us to represent perigee states,
and secondly it doesn't map directly onto the outlier types we need. In
addition, the TrackState types are converted to TrackStateOnSurfaceType
for the actual creation of the track. It turns out that using
TrackStateOnSurfaceType is actually a rather more elegant way of
capturing the track state types in our internal representation.  This
commit changes the use of TrackStateType in the GXFTrackState to
TrackStateOnSurfaceType.