drop unused tau calibration
Hello,

This MR drops a tau energy calibration that is not used anywhere in the tau reconstruction (the PanTauCellBasedProto calibration).
It prepares the ground for upcoming tau EDM revisions (ATLTAU-1720), where this calibration will be dropped.
I've checked that the only change in the tau reconstruction output is the TauJetsAux.ptPanTauCellBasedProto,etaPanTauCellBasedProto,phiPanTauCellBasedProto.mPanTauCellBasedProto variables, which are now 0.

Cheers,
Bertrand