Store perigee in internal trajectory format
Currently, perigee parameters are not stored in the GlobalChi2Fitter
internal track representation. The reason for this might have been that
the architecture of that internal representation does not support
perigee parameters, but this has since been changed. The fact that
perigee parameters are not stored is a problem if we need to perform
operations on entire tracks in intermediate representations as the
perigee might be missing. This commit changes the track creation
process such that the perigee is added to the track before converting
it to the Trk::Track representation.