Reformat GlobalChi2Fitter CMakeLists.txt
The existing formatting of this file is somewhat messy and has some
extremely long lines which are not very readable to humans and also
confusing for version control systems, as the entire line needs to be
changed to modify one single element of the list. This commit changes
the formatting of the CMakeLists.txt file to be a lot neater and easier
to process for VCS. The actual contents of the file are not changed,
and the end result should be identical.