drop unused tau variable
Hello,

This MR is dropping a variable that is not used anywhere (anymore) in the tau reconstruction.
We now use a Et-resolution weighted average between calo and substructure pt as input to the MVA pt, no longer an interpolation.
I have checked it does not change the reconstruction output, besides this variable that is now gone.
Minor code cleanup was done.

Cheers,
Bertrand