Add another workaround the TDT feature retrieval matching "feature"
The TDT returns features with a "matching" key as well as the with the requested key.
if a matching keyed object is found first, then the object with the requested key
cannot be retieved as "lastOfType", so all such features need to be requested.

When no providing a key, then requesting all features returns all the features,
instead of just the last of type, so these two cases - with a key, and without
a key have to be treated differently.

This change will be able to be reoved once the TDT bug is fixed, hence the lines
have been left in, but commented.