tauRecTools: fix small bug in dR calculation (TauCommonCalcVars)
Hello,

This MR fixes a small bug in a dR calculation, which uses eta and phi from inconsistent tau calibrations (it only affects the offline reconstruction, in the trigger it is fine).
The fix should have limited impact on the reconstruction output, as it's a vertex correction in phi which is not picked up, which should be small (e.g. compared to a vertex correction in eta).
Minor cleanup done elsewhere.

Cheers,
Bertrand