Use unique_ptr for tracks in MuonTrackSteeringTools
Switching to using unique_ptr for Track and MuPatTrack objects.  In particular, MuPatTrack now stores a unique_ptr to the Track, rather than an (unnecessarily) const bare pointer.
This should simplify the memory management for tracks and eliminates the need for some const_casts.
Additionally, MuonTrackSelectorTool is moved to MuonTrackSteeringTools, as it is only used there.