First pass at overlay setup
Much of our code relies on DerivationFrameworkIsMonteCarlo , which is
actually not what it means to use. This flag was set based on the
"input" being "geant4". For overlay, the input is set to data. What most
of these pieces of code actually want to know is whether there is truth
information in the input file. Therefore, splitting off a clear
"HasTruth" flag, including a separate "HasxAODTruth" flag. Migrating all
clients I could identify.

This also brought to light a few places where we use MC / truth
information as a proxy for something else (e.g. an unprescaled or
different trigger menu). It'd be better if those pieces of code could be
migrated to check what they actually intend to look for, but that's
beyond the scope of this MR. Here I tried to add some comments to point
those places out.

This should be a first step towards solving ATLASG-1532.