tauRecTools: drop pileup-corrected tau ID variable
Hello,

This MR is dropping a couple of pileup-corrected tau ID variables which are no longer used.
There was also a decoration of variable used nowhere (duplicate decoration).
I also changed all the SG::AuxElement to static const, following a remark in a previous MR, but I have not checked the impact on the performance.

Cheers,
Bertrand