Repairing truth tau building for overlay in 21.2
This MR fixes the next issue in ATLASG-1532, the building of truth taus
during jobs running on overlay events. Here the metadata indicates that
the job is data, but the file has truth info in it, so we needed an
override in the case that we've identified that we're running on
overlay.