tauRec: drop TauCommonPi0Cells container from xAOD
Hello,

This MR is dropping the TauCommonPi0Cells container, it is too large to afford it in xAODs (it was not written out in R21 xAODs).
We will write it out on demand for performance samples.
Note to ourselves: this means we can no longer rerun our custom topoclustering at xAOD level, so we need to double check it is well configured in R22:
https://gitlab.cern.ch/atlas/athena/blob/master/Reconstruction/tauRec/share/Pi0ClusterMaker_jobOptions.py

Bright side: we save about 20 kb per event!!

Cheers,
Bertrand