PyUtils: Add batch option to ROOT6Setup
With the new PyROOT the command line argument `-b` is no longer parsed
to set the batch mode. Add an option to `ROOT6Setup` to set batch mode
whenever the first `ROOT` import is done. This can then be used by
`athena.py` to set the batch mode but without having to actually import
ROOT early in the boostrap.