RingerSelectorTools: Use PATCoreAccept instead of PATCore
RingerSelectorToolsEnumDict is loaded during configuration, and cannot
depend on Athena libraries or we may see failures with loading converters.
It depends on PATCore, which used to have minimal dependencies, but now
depends on AthenaBaseComps.

The pieces we need from PATCore were split into a separate library,
so depend on that instead.

Also clean up the cmake file.