Avoid recomputation of holes in TrackSummaryTool
As of !36404, the TrackSummaryTool no longer ignores tracks with
existing track summaries. However, this has introduced a bit of a
performance regression where it recomputes hole searches for tracks
that already have that infomation attached. This commit adds a
constraint to the hole search code such that it only runs if the track
does not already have hole search information attached to it. As far as
I can tell this solves the performance bug.