Fix for ATLASRECTS-5661
Using the dynamic ID alignment leads to rare crashes in when calculating the etcone isolation for muons.
This turns out to be because extrapolation into the calorimeter from the ID exit fails.
During standard reco, extrapolation into the calorimeter is done using the perigee: as it turns out, this works with dynamic ID alignment as well.
Therefore this fix updates the ParticleCaloCellAssociationTool to ensure that its instance of the ParticleCaloExtensionTool always starts from the perigee.
The MuonCreatorTool is also updated, as the association of cells to muon to form a cluster will fail otherwise (although this does not lead to a crash).
In principle, for muons the difference between starting the extrapolation at the ID exit and the perigee should be negligible.
While it would be good to understand exactly why the use of the alignment causes this crash, this fix will be much simpler and quicker to implement.