AthenaConfiguration: Fix confTool behaviour for old configurations.
Previously, when reading old configurations, confTool would merge
together the two dictionaries that were written.  This is needed
for proper handling of JobOptionsSvc; the properties we want to compare
are those from the second dictionary.  (The first dictionary
contains properties such as the search path which we want to ignore.)
This was changed when confTool was unified for the old and new
configurations.  Tweak it to get back to the previous behavior.