Switch back to default pixel cluster tool for MuonRotCreator
Reverting one of the changes from https://gitlab.cern.ch/atlas/athena/-/merge_requests/36431 which updated the pixel cluster tool used by default in the MuonRotCreator. This tool is not actually used (it's disabled in the muon 'mode'), so it shouldn't really matter which tool is configured here. However, it seems that as it is now is somehow configuring some tracking tools out of place in some offline and trigger jobs (maybe the import is picking up something not wanted in this particular case). See also ATLASRECTS-5664.