tauRecTools: use consistent treatments of vertex correction (ATLTAU-1716)
The vertex correction is not applied consistently in the tau
reconstruction. In addition, it is applied to the four momentum already
corrected by PV0, indicating an over correction.

Ideally, we could apply the vertex correction to the four momentum
before the origin correction in jet reconstruction. But it is not
straight forward to obtain this information. Alternatively, we
apply the vertex correction relative to the jet vertex to fix
this bug.