AthAnalysisBaseComps: Fix AthAnalysisHelper::setProperty.
AthAnalysisHelper::setProperty was changed from
IJobOptionsSvc::addPropertyToCatalogue to IOptionsSvc::set.
However, the functionality is not identical.
When settingg a string property, the old interface would strip off
quotes surrounding the value; the new interface does not do this.
Further, the code in AthAnalysisHelper would uncoditionally use
Gaudi::Utils;:toString, which adds quotes around values
which are strings.

So if we have something convertible to a string, don't use toString().