Update rois used for run 2 muon triggers
For the Run 3 triggers, we use the RoIs created by the L2 algorithms directly for the precision muon finding which have some variation in the eta/phi widths based on position. For the Run 2 triggers, we were taking the positions from L2, but using a fixed size regardless of the detector position. Updating the Run 2 code to use the eta/phi widths from the L2 RoIs as well since it's a more sensible approach (and to make comparisons with the Run 3 triggers easier).