Fixed the tests of PhotonVertexSelection for Athena and AnalysisBase.
Since the package was not built as part of Athena before, it is now
added to that project as well.

The test in Athena doesn't actually run anything... It sets up a job,
but doesn't run any event loop. However I didn't want to completely
re-design the test to set things up correctly for an Athena test.
(The Athena and standalone tests should just be done using separate
Python files, as they don't actually share any meaningful amount of
code with each other.)