remove empty file with space in file name
This file is empty, and worse it contains a space in the name which
messes up `find` commands running over all files in the repository.
From what I can tell it has come straight from SVN years ago and not
been touched since.  So hopefully just removing this file is fine, and
doesn't cause other things to break that depend on its mere existence.