Fixes to allow the low pt analyses to be run in the new boilerplate free code
Many of the tests, have very similar configurations, and the code is all duplicated.

These changes factorise the common code better, and allows different analyses
for the low pt chains in the common configuartion baseclass.

Renames the output directory for the offline analyses also to reduce the need for
duplicated flags.