tauRec: create pi0 cell container on demand
Hello,

This MR is making one more step towards the tau re-reconstruction from xAOD (ATLTAU-1733).
It introduces the possibility not to create the pi0 cell container (TauCommonPi0Cells) when rerunning from xAOD.
This relies on the fact that TauCommonPi0Cells is available in xAODs, which is not longer the case since !36398 .
It is now put back in the xAOD. In a MR to follow shortly, we will thin TauCommonPi0Cells to only keep the cells within the pi0 topoclusters, which will reduce its size by a factor of 5.
TauRecBuilder was also cleaned up.

Cheers,
Bertrand