tauRecTools: adapt TauVertexVariables to rerun on xAOD
Hello,

This MR is preparing the ground for an upcoming MR that will allow the full tau reconstruction to be rerun at xAOD level (ATLTAU-1733).
In tauRecTools, the only tool that needs to be modified is TauVertexVariables.
As the navigation from xAOD::TrackParticle to Trk::Track is not possible in xAODs, we need to use a std::vector<const Trk::TrackParameters*> as input to the VertexSeedFinder, instead of a std::vector<const Trk::Track*>.
https://gitlab.cern.ch/atlas/athena/-/blob/master/Tracking/TrkVertexFitter/TrkVertexSeedFinderTools/src/CrossDistancesSeedFinder.cxx#L89
https://gitlab.cern.ch/atlas/athena/-/blob/master/Tracking/TrkVertexFitter/TrkVertexSeedFinderTools/src/CrossDistancesSeedFinder.cxx#L112

I have verified that the 2 methods give the consistent results (up to rounding effects) using standalone tau reconstruction from ESDs, where Trk::Track is still available.

All the remaining changes will happen in tauRec, not tauRecTools, and will be implemented in a follow-up MR.

Cheers,
Bertrand