Adding muComb to trigger new JO test
Adding the L2 combined muon step to the trigger new JO test (ATR-20242). The step runs:
- A view data verifier to pick up the L2 SA muons from the previous step
- The ID fast track finding
- the muComb algorithm

Currently it seems that the combined electron+muon chains need to have the same number of steps for electrons and muons, so I split the muon chains into two type. The default chains (used in combination with other signatures) run the L2 SA and combined steps, and the msonly chains run the L2 SA and precision SA steps (and the L2 combined and preicion SA steps are only created for the relevant chains).