Add hole search logic to the GlobalChi2Fitter
This commit adds the necessary logic to do hole search to the
GlobalChi2Fitter. The implementation is very close in output to the
InDetTrackHoleSearch, and may even improve on its quality slightly. The
hole search, if enabled, adds some CPU time to the fitting process but
this is negated by reduced CPU time later on in the algorithm chain. The
hole search data is carried out of the fitter through a track summary
object attached to the track. This functionality is disabled by default
to preserve exact bit-to-bit equality of the output, but can be enabled
using the DoHoleSearch flag. The rest of Athena should by now be
configured to accept this and work around it without performing
unnecessary additional hole searches.