VarHandleKey: Support for DataHandle Python type
Implement Python type support for handles. All [Read,Write,Cond]Handles
are represented as `GaudiKernel.DataHandle`. The following changes were
done:
1) Override `VarHandleKey::pythonRepr` to write the proper Python
representation during the Configurable generation.
2) Make `VarHandleKeyProperty` derive from `DataHandleProperty`. This is
necessary in order for the property to be recognized as (data) handle in
`genconf`.

A few unit tests had to be adapted as the property type is now correctly
propagated by Gaudi. E.g. a `WriteHandleKey` property reports its type
whereas it was `VarHandleKey` before. Note that `WriteDecorHandleKey`
reports now as `WriteHandleKey` as well because our implementation
leverages the `declareProperty` of its base.

Relates to gaudi/Gaudi!1091.