copy over PileupReweighting from 21.2
I had started out by sweeping changes from 21.2 to master, but that
turned out to be a big mess, given that a lot of changes were applied
in either branch and many in both.  So this is now just removing the
package as it was in master and copying over the package as it is in
21.2 right now, followed by some cleanup of obvious places where
development work was undone or where what was in 21.2 was clearly out
of line with current best practices.