Remaining changes to remove the RegSelSvc ID use
These changes (hopefully) finish the change needed to remove the use of the RegSelSvc
for the ID sub-systems.

In addition, there are changes to the RegSelSvc code and the IDRegionSelector
code to remove the generation of the lookup tables using the old code, and remove
the RS functionality for the Pixel, SCT and TRT from the old RegSe;Svc.

In addition, to get round an issue with the PixelCablingCondAlg which seems to
report an unusable EventIDRange, there is a hack to get it to return a range
which should (mostly) work.

This hack should be replaced by a proper impementation of the PixelCablingCondAlg
which no longer needs to set up during the initialise (at least for the RegionSelector)

If the pixel cabling still needs to be configured during initialize() for some
other reason, then this range should be fixed such that it provides a valid range
for the subseuqnet use, or overwridden properly in the execute method.