Remove the default boundary tool from GX2 fitter
Previously, the GlobalChi2Fitter can equipped with a default value for
its boundary checking tool tool handle. However, I now believe that this
might be causing some CI test failures, so to be safe I am removing this
default value from the fitter. It now comes with an empty default value,
and the hole search functionality is automatically disabled if no tool
is provided.