Add a CxxUtils type for arrays of optionals
This commit adds a type to the CxxUtils package, the OptionalArray,
which for a given template type T wraps the std::array<std::optional<T>>
type. It adds several useful helper methods which make the type more
ergonomic to use. This can be used later in applications such as the
Trk::TrackSummary, or other places where an array of optional values is
required.