Use Eigen types in transformPlaneToGlobal
Currently, the transformPlaneToGlobal method in the Runge Kutta
utilities package uses bare arrays of doubles. This means that data
needs to be copied out of AMG (Eigen) vectors in order to use them
there. This commit changes the signature of this function to take
AmgVector(5) references as argument instead, which should reduce data
copying a little, and bring the method more in line with the design of
the track parameter classes.