Use Eigen vector for pattern track parameters
This is the first of several changes to align the data member types of
the PatternTrackParameters class with those of the ParametersBase class,
so that we can eventually unify these currently distinct types under one
type hierarchy. We start out by changing the actual parameters (a
pentuple) to be stored in an Eigen vector, as is the case for the
ParametersBase.