Implement Eigen pattern parameter vector getter
The existing getter for the parameter vector in the pattern track
parameters returns a pointer to a floating point value, reminiscent of
old-school C-type array passing. This is not compatible with the getter
signatures for the ParametersBase, where a reference to an Eigen vector
is returned. In the previous commit we ensured that thee pattern track
parameters actually contained an Eigen vector for the parameters, which
we can now simply return. Of course, this also means we must change the
use of the getter method in other parts of the code.